asteroid/requirements.txt
django
django-extensions
amqplib
django-guardian